+++ tests/hash_tests.rs
@@ -1,5 +1,3 @@
-use std::alloc::{alloc, Layout};
-
 use digest::Digest;
 use hex_literal::hex;
 
@@ -35,14 +33,11 @@ fn validate_samples() {
 
 #[test]
 fn validate_with_buffer() {
-    let mut scratchpad = unsafe {
-        let buffer = alloc(Layout::from_size_align_unchecked(CryptoNight::SP_SIZE, CryptoNight::SP_ALIGNMENT));
-        Vec::from_raw_parts(buffer, CryptoNight::SP_SIZE, CryptoNight::SP_SIZE)
-    };
+    let mut scratchpad = CryptoNight::allocate_scratchpad();
 
     for (i, (&input, &output)) in INPUTS.iter().zip(OUTPUTS.iter()).enumerate() {
         println!("{}: {}", i, hex::encode(input));
-        let result = CryptoNight::digest_with_buffer(input, &mut scratchpad);
+        let result = CryptoNight::digest_with_buffer(input, scratchpad.as_mut());
 
         assert_eq!(result[..], output[..]);
     }